Extract submitLog helper
