+++ packages/plugins/analytics/src/constants.ts
@@ -5,5 +5,3 @@
 import type { PluginName } from '@dd/core/types';
 
 export const PLUGIN_NAME: PluginName = 'datadog-analytics-plugin' as const;
-export const INTAKE_PATH = 'v1/input/pub44d5f4eb86e1392037b7501f7adc540e';
-export const INTAKE_HOST = 'browser-http-intake.logs.datadoghq.com';
