+++ packages/plugins/analytics/src/index.ts
@@ -2,10 +2,10 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-Present Datadog, Inc.
 
-import { doRequest } from '@dd/core/helpers/request';
+import { submitLog } from '@dd/core/helpers/log';
 import type { GetInternalPlugins, GetPluginsArg } from '@dd/core/types';
 
-import { INTAKE_HOST, INTAKE_PATH, PLUGIN_NAME } from './constants';
+import { PLUGIN_NAME } from './constants';
 
 export { PLUGIN_NAME } from './constants';
 
@@ -20,39 +20,7 @@ export const getAnalyticsPlugins: GetInternalPlugins = (arg: GetPluginsArg) => {
         }
 
         try {
-            const bundler = {
-                name: context.bundler.name,
-                version: context.bundler.version,
-            };
-
-            await doRequest({
-                // Don't delay the build too much on error.
-                retries: 2,
-                minTimeout: 100,
-                url: `https://${INTAKE_HOST}/${INTAKE_PATH}`,
-                method: 'POST',
-                type: 'json',
-                getData: async () => {
-                    const data = {
-                        ddsource: `@datadog/${bundler.name}-plugin`,
-                        env: context.env,
-                        message,
-                        service: 'build-plugins',
-                        bundler,
-                        metadata: context.build.metadata,
-                        plugins: context.pluginNames,
-                        version: context.version,
-                        team: 'language-foundations',
-                        ...overrides,
-                    };
-                    return {
-                        data: JSON.stringify(data),
-                        headers: {
-                            'Content-Type': 'application/json',
-                        },
-                    };
-                },
-            });
+            await submitLog({ context, message, rest: overrides });
         } catch (e: unknown) {
             // We don't want to break anything in case of error.
             log.debug(`Could not submit data to Datadog: ${e}`);
